Rename debug protocol interfaces
